fix Postgres-related code

+++ sb2/src/main/resources/db/migration/V2.0__location.sql
@@ -1,7 +1,7 @@
 create table location (
     id        serial       not null
         constraint location_pkey primary key,
-    zip       int,
+    zip       bigint,
     city      varchar(250) not null,
     person_fk serial
         constraint person_fk references person
